requirements.txt
pandas==1.5.3
numpy==1.25.2
scikit-image==0.19.3
matplotlib==3.7.1
mat73==0.62
jupyter==1.0.0
dvc[gdrive]
pyyaml